+++ src/components/comments/comments.jsx
@@ -7,8 +7,8 @@ import TreeView from "@material-ui/lab/TreeView";
 import TreeItem from "@material-ui/lab/TreeItem";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-import useComments from './useComments';
-import { textParser } from '../../utils/utils';
+import useComments from './use-comments';
+import { textParser } from '../../shared/utils/utils';
 import Loader from '../loader/loader';
 import Error from '../error/error';
 import classes from './comments.module.css';
