+++ src/components/single-news-item/use-news-item.js
@@ -1,17 +1,18 @@
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router";
 import { loadSingleNews } from "../../redux/actions/single-news";
-import { newsItemSelector } from "../../redux/selectors/news-selector";
 
 export default () => {
   const { id } = useParams();
   const dispatch = useDispatch();
-  const [loading, setLoading] = useState(false);
   
-  const { newsItems, isItemsFailed } = useSelector(newsItemSelector);
+  const { newsItems, isItemsFailed } = useSelector(state => state.newsItemReducer);
   const singleItem = newsItems[id];
-  const { item } = singleItem || {};
+  const { item, isLoaded } = singleItem || {};
+
+  console.log(`isLoaded`, isLoaded);
+
   
   useEffect(() => {
     if (singleItem) return;
@@ -20,15 +21,11 @@ export default () => {
 
   const requestSingleNews = async () => {
     try {
-      setLoading(true);
       dispatch(loadSingleNews(id));
-      setLoading(false);
     } catch(e) {
       console.error(`Что-то пошло не так: ${e}`);
     }
-  };
+  };  
 
-  
-
-  return { item, loading, isItemsFailed };
+  return { item, isLoaded, isItemsFailed };
 };
