+++ src/shared/utils/comments/comments.js
@@ -1,3 +1,5 @@
+import { baseURL } from "../../requests/api";
+
 export const buildTree = (comments, postId) => {
   const commentsById = {};
